fix testing with test project
since we cant simulate a second .git folder in the tests/data directory we use the one at the project root. This means we can't point the Experiment project_path variable at the test/data directory where the Dockerfile etc is. This commit gets around that by only running check_dirty when the Experiment is not. being run by pytest